% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_procesamiento.R
\name{procesar_prop_voto_sen_tod_op}
\alias{procesar_prop_voto_sen_tod_op}
\title{Title}
\usage{
procesar_prop_voto_sen_tod_op(datos_recibidos)
}
\arguments{
\item{datos_recibidos}{}
}
\description{
Title
}
